Use fake zones for test
